refactor: edit index (#61)
