+++ src/hooks/index.ts
@@ -1,1 +1,3 @@
 //hooks
+
+export { useCommentSort } from './useCommentSort'
