+++ src/utils/index.ts
@@ -3,3 +3,4 @@
 export { cn } from './cn'
 export { formatTimeAgo } from './formatTimeAgo'
 export { getPageNumbers } from './getPageNumbers'
+export { formatRelativeTime } from './formatRelativeTime'
